requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-multipart==0.0.9
pandas==2.2.0
pydantic==2.6.1
numpy==1.26.3
scikit-learn==1.4.0
pdfplumber==0.10.3
openai==1.12.0
langchain==0.1.4
passlib[bcrypt]==1.7.4
python-jwt==4.0.0
plaid-python==18.0.0
pydantic-settings==2.1.0
